remove site from sync manifest endpoint
