+++ server/apps/sites/tasks/__init__.py
@@ -1,2 +1,2 @@
 from .profile_tasks import sync_user_sites
-from .site_tasks import get_rcra_site, sync_site_manifests
+from .site_tasks import get_rcra_site
