+++ server/apps/trak/views/__init__.py
@@ -1,3 +1,9 @@
 from .handler_view import HandlerView, TransporterView
 from .lookup_views import FederalWasteCodes
-from .manifest_view import ManifestView, MtnList, PullManifestView, SignManifestView
+from .manifest_view import (
+    ManifestView,
+    MtnList,
+    PullManifestView,
+    SignManifestView,
+    SyncSiteManifestView,
+)
